Finalise new build system
The new build system is miles better than the old one. By splitting out standalone programs and libraries (into src and lib respectively) it's now possible to generate a rule for each program. If a program needs a library, it can be listed in a variable called `program_name_deps`, but if it doesn't then it is automatically built.  There is automatic dependency checking, so if a file changes then everything effected by it can be rebuilt with the command `make`. 

The workflow now is: make a change, and then either type `make` to rebuild everything that needs to be, `make program_name` to build the just the one program (often these commands will be equivalent, so it's faster to type make), or type `make program_name.install` to build and install the specific program you changed. Because the makefile uses static patterns, tab completion just works.  

Each file in `src` is treated as a separate, installable program for the mbed, i.e. it's expected to have a `main()` function. If the program needs to be split into other files, put them into `lib'.

Sub-directories in `lib` get compiled into static .a files (static libraries). This means a collection of libraries, possibly with inter-dependencies, can be added as one file in the `program_name_deps` variable. See, for example, the ioboard library, which can be linked with a program by adding `ioboard.a` to the list.

As a final note, the binaries have not been tested as of yet on hardware, so there may be regressions.